fix: add support for building in non-git source dir
The git_version cmake macro uses git tags to determine the library
version number. However, if the source directory does not have a .git
repo, the cmake configuration step fails.

The git_versino macro will now look for a file named version.txt in the
root directory if the source dir is not a git repository. If the file
exists, its contents are used for the version number instead of git
describe.

Now we can have the tag-release.sh script update version.txt when
tagging new releases.